Remove StaticCharAtlas support (fixes #1576)
This defaults to the dynamic atlas and leaves in NoneCharAtlas and the
experimentalCharAtlas option (with 'static' removed).

This leaves me with a few questions:

1. Do we want to completely remove support for NoneCharAtlas and
   experimentalCharAtlas? NoneCharAtlas provides a nice example of the
   modularity of the atlas system, even though it doesn't make much
   sense to use.

2. If we want remove the experimentalCharAtlas option, should we leave
   the typedef there (ignoring any values passed) for backwards
   compatibility? I think semver requires this, though it's not a
   runtime concern.

3. If we want to keep experimentalCharAtlas, should we rename it to
   charAtlas? What should we do if someone passes in an
   experimentalCharAtlas option? Should we leave 'static' in there as an
   alias to 'dynamic' for compatibility? I think semver requires this.

4. Unfortunately I don't think we have any established standard for
   warning about deprecated features. Is it appropriate to use
   `console.warn` in a library context like this? My guess is no.
