+++ src/renderer/atlas/CharAtlasCache.ts
@@ -10,11 +10,9 @@ import { generateConfig, configEquals } from './CharAtlasUtils';
 import BaseCharAtlas from './BaseCharAtlas';
 import DynamicCharAtlas from './DynamicCharAtlas';
 import NoneCharAtlas from './NoneCharAtlas';
-import StaticCharAtlas from './StaticCharAtlas';
 
 const charAtlasImplementations = {
   'none': NoneCharAtlas,
-  'static': StaticCharAtlas,
   'dynamic': DynamicCharAtlas
 };
