+++ typings/xterm.d.ts
@@ -77,17 +77,10 @@ declare module 'xterm' {
      * artifacts.
      *
      * - 'none': Don't use an atlas.
-     * - 'static': Generate an atlas when the terminal starts or is reconfigured. This atlas will
-     *   only contain ASCII characters in 16 colors.
      * - 'dynamic': Generate an atlas using a LRU cache as characters are requested. Limited to
-     *   ASCII characters (for now), but supports 256 colors. For characters covered by the static
-     *   cache, it's slightly slower in comparison, since there's more overhead involved in
-     *   managing the cache.
-     *
-     * Currently defaults to 'static'. This option may be removed in the future. If it is, passed
-     * parameters will be ignored.
+     *   ASCII characters (for now), but supports 256 colors. (recommended)
      */
-    experimentalCharAtlas?: 'none' | 'static' | 'dynamic';
+    experimentalCharAtlas?: 'none' | 'dynamic';
 
     /**
      * The font size used to render text.
@@ -474,7 +467,7 @@ declare module 'xterm' {
      * (EXPERIMENTAL) Registers a character joiner, allowing custom sequences of
      * characters to be rendered as a single unit. This is useful in particular
      * for rendering ligatures and graphemes, among other things.
-     * 
+     *
      * Each registered character joiner is called with a string of text
      * representing a portion of a line in the terminal that can be rendered as
      * a single unit. The joiner must return a sorted array, where each entry is
@@ -483,16 +476,16 @@ declare module 'xterm' {
      * a single unit. When multiple joiners are provided, the results of each
      * are collected. If there are any overlapping substrings between them, they
      * are combined into one larger unit that is drawn together.
-     * 
+     *
      * All character joiners that are registered get called every time a line is
      * rendered in the terminal, so it is essential for the handler function to
      * run as quickly as possible to avoid slowdowns when rendering. Similarly,
      * joiners should strive to return the smallest possible substrings to
      * render together, since they aren't drawn as optimally as individual
      * characters.
-     * 
+     *
      * NOTE: character joiners are only used by the canvas renderer.
-     * 
+     *
      * @param handler The function that determines character joins. It is called
      * with a string of text that is eligible for joining and returns an array
      * where each entry is an array containing the start (inclusive) and end
